+++ src/main/scala/cromwell/engine/package.scala
@@ -3,8 +3,8 @@ package cromwell
 import java.nio.file.{Path, Paths}
 
 import cromwell.engine.io.gcs.GcsFileSystem
+import org.apache.commons.lang3.exception.ExceptionUtils
 import org.joda.time.DateTime
-import org.slf4j.Logger
 import wdl4s._
 import wdl4s.values.{WdlFile, WdlValue}
 
@@ -27,6 +27,7 @@ package object engine {
 
   final case class ExecutionEventEntry(description: String, startTime: DateTime, endTime: DateTime)
   final case class ExecutionHash(overallHash: String, dockerHash: Option[String])
+  final case class CallAttempt(fqn: FullyQualifiedName, attempt: Int)
 
   type ErrorOr[+A] = ValidationNel[String, A]
 
@@ -44,6 +45,7 @@ package object engine {
   type HostInputs = Map[String, WdlValue]
 
   class CromwellFatalException(exception: Throwable) extends Exception(exception)
+  class PreemptedException(msg: String) extends Exception(msg)
 
   implicit class EnhancedFullyQualifiedName(val fqn: FullyQualifiedName) extends AnyVal {
     def isScatter = fqn.contains(Scatter.FQNIdentifier)
@@ -65,10 +67,10 @@ package object engine {
 
       def isUriWithProtocol: Boolean = "^[a-z]+://".r.findFirstIn(str).nonEmpty
 
-      def toPath(workflowLogger: Logger, gcsFileSystem: Try[GcsFileSystem] = Failure(new Throwable("No GCS Filesystem"))): Path = {
+      def toPath(gcsFileSystem: Try[GcsFileSystem] = Failure(new Throwable("No GCS Filesystem"))): Path = {
         str match {
           case path if path.isGcsUrl && gcsFileSystem.isSuccess => gcsFileSystem.get.getPath(str)
-          case path if path.isGcsUrl => throw new Throwable(s"Unable to parse GCS path $path: ${gcsFileSystem.failed.get.getMessage}")
+          case path if path.isGcsUrl => throw new Throwable(s"Unable to parse GCS path $path: ${gcsFileSystem.failed.get.getMessage}\n${ExceptionUtils.getStackTrace(gcsFileSystem.failed.get)}")
           case path if !path.isUriWithProtocol => Paths.get(path)
           case path => throw new Throwable(s"Unable to parse $path")
         }